refactor(phone): one thread, drawn and answered the same way everywhere
The threads screen owned all of it: the card, the reply box, the resolve
button, the confirmation before a suggestion is committed, and the three
writes underneath them. That was right while it was the only place a
conversation could be answered. The diff is about to draw the same
threads on the lines they are about, and two copies of "reply, resolve,
apply" is how one of them quietly stops re-reading after a write.

So the card is a component, the confirmation is a component, and the
writes are a hook the caller hands its own re-read to — the two screens
read different things and both must re-read.

Nothing changes on screen except one thing worth having: a comment body
now goes through the markdown renderer. A review comment is where a
suggestion block lives, and a ```suggestion fence printed verbatim was
the worst thing this app showed — five lines of markup wrapped at eleven
points with the proposed code hidden inside them.

The screen is 381 lines to 117, and what is left in it is the list.
